fix(pdf): add 'pdfjs' element and restore native iframe PDF viewer as default
