+++ backend/chainlit/__init__.py
@@ -31,6 +31,7 @@ from chainlit.element import (
     File,
     Image,
     Pdf,
+    Pdfjs,
     Plotly,
     Pyplot,
     Task,
@@ -164,6 +165,7 @@ __all__ = [
     "ModeOption",
     "OutputAudioChunk",
     "Pdf",
+    "Pdfjs",
     "PersistedUser",
     "Plotly",
     "Pyplot",
